apps/api/requirements_admin.txt
# FastAPI Requirements

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.8.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dateutil==2.8.2
email-validator==2.1.0
requests==2.31.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
